+++ .storybook/theme.js
@@ -1,5 +1,5 @@
 import { create } from '@storybook/theming/create';
-import ReagraphLogo from '../docs/assets/logo.svg';
+import ReagraphLogo from '../stories/assets/logo.svg';
 
 export default create({
   base: 'dark',
